Move ProfileDatabase into src to simplify distributed includes.
