+++ configure.ac
@@ -261,8 +261,8 @@ AC_CONFIG_LINKS([include/souffle/CompiledRecord.h:src/CompiledRecord.h])
 AC_CONFIG_LINKS([include/souffle/CompiledSouffle.h:src/CompiledSouffle.h])
 AC_CONFIG_LINKS([include/souffle/SouffleInterface.h:src/SouffleInterface.h])
 AC_CONFIG_LINKS([include/souffle/ParallelUtils.h:src/ParallelUtils.h])
-AC_CONFIG_LINKS([include/souffle/profilerlib/ProfileDatabase.h:src/profilerlib/ProfileDatabase.h])
-AC_CONFIG_LINKS([include/souffle/profilerlib/json11.h:src/profilerlib/json11.h])
+AC_CONFIG_LINKS([include/souffle/ProfileDatabase.h:src/ProfileDatabase.h])
+AC_CONFIG_LINKS([include/souffle/json11.h:src/json11.h])
 AC_CONFIG_LINKS([include/souffle/ProfileEvent.h:src/ProfileEvent.h])
 AC_CONFIG_LINKS([include/souffle/RamTypes.h:src/RamTypes.h])
 AC_CONFIG_LINKS([include/souffle/AstTypes.h:src/AstTypes.h])
